requirements.txt
weaviate-client==4.6.3
